pyproject.toml
[project]
name = "mytaskie"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.12",
    "google-api-python-client>=2.166.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "google-genai>=1.10.0",
    "jinja2>=3.1.6",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.0",
    "pytz>=2025.2",
    "redis>=5.2.1",
    "tzlocal>=5.3.1",
    "urllib3==1.26.6",
    "uvicorn>=0.34.0",
]
